spec(replacePathVars): add tests
